Fixed Bazel issues and cleaned up IntelliJ files
