--- .idea/hydra.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="HydraSettings">
-    <option name="hydraStorePath" value="$PROJECT_DIR$/.hydra/idea" />
-    <option name="noOfCores" value="4" />
-    <option name="projectRoot" value="$PROJECT_DIR$" />
-    <option name="sourcePartitioner" value="auto" />
-  </component>
-</project>
