--- .idea/modules.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/project_root.iml" filepath="$PROJECT_DIR$/project_root.iml" />
-      <module fileurl="file://$PROJECT_DIR$/src/main/src_main.iml" filepath="$PROJECT_DIR$/src/main/src_main.iml" group="hpc" />
-    </modules>
-  </component>
-</project>
