+++ BUILD
@@ -13,7 +13,7 @@ deps = require(
     "@com_fasterxml_jackson_core_jackson_core",
     "@com_fasterxml_jackson_core_jackson_databind",
     "@com_fasterxml_jackson_core_jackson_databind_yaml",
-    "@edu_stanford_nlp_standford_corenlp",
+    "@edu_stanford_nlp_stanford_corenlp",
     "@org_apache_commons_commons_lang3",
     "@org_yaml_snakeyaml",
 )
