--- project_root.iml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$">
-      <excludeFolder url="file://$MODULE_DIR$/.buckd" isTestSource="false" />
-      <excludeFolder url="file://$MODULE_DIR$/.git" isTestSource="false" />
-      <excludeFolder url="file://$MODULE_DIR$/buck-out" isTestSource="false" />
-      <excludeFolder url="file://$MODULE_DIR$/resources" isTestSource="false" />
-    </content>
-    <orderEntry type="jdk" jdkName="1.8" jdkType="JavaSDK" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="//:scala-compiler-lib" scope="COMPILE" level="project" />
-    <orderEntry type="module" module-name="src_main" scope="COMPILE" />
-  </component>
-</module>
