+++ src/main/scala/com/linguistic/patterson/agents/BUILD
@@ -3,7 +3,9 @@ scala_library(
     srcs = glob(["*.scala"]),
     visibility = ["//visibility:public"],
     deps = require(
+        "@edu_stanford_nlp_stanford_corenlp",
         "clients",
+        "models/local",
         "models/corenlp",
         "util"
     )
