+++ src/main/scala/com/linguistic/patterson/models/corenlp/BUILD
@@ -2,7 +2,5 @@ scala_library(
     name = "corenlp",
     srcs = glob(["*.scala"]),
     visibility = ["//visibility:public"],
-    deps = require(
-        "@edu_stanford_nlp_standford_corenlp",
-    ),
+    deps = require("@edu_stanford_nlp_stanford_corenlp"),
 )
