--- src/main/src_main.iml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/scala" isTestSource="false" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="//:commons-lang3" level="project" />
-    <orderEntry type="library" name="//:corenlp" level="project" />
-    <orderEntry type="library" name="//:jackson-annotations" level="project" />
-    <orderEntry type="library" name="//:jackson-core" level="project" />
-    <orderEntry type="library" name="//:jackson-databind" level="project" />
-    <orderEntry type="library" name="//:jackson-dataformat-yaml" level="project" />
-    <orderEntry type="module" module-name="project_root" />
-    <orderEntry type="library" name="//:scala-library" level="project" />
-    <orderEntry type="library" name="//:snakeyaml" level="project" />
-    <orderEntry type="library" name="scala-sdk-2.12.2" level="application" />
-  </component>
-</module>
